+++ examples/calculator/src/contract.rs
@@ -1,37 +1,37 @@
 use crate::error::CalcError;
-use crate::message::CalcMsg;
-use kelk_env::{context::Context, Response};
+use crate::message::{ProcMsg, QueryMsg, QueryRsp};
+use kelk_env::context::Context;
 
-fn add(_ctx: Context, a: i32, b: i32) -> Result<i32, CalcError> {
-    Ok(a + b)
+fn add(ctx: Context, a: i32, b: i32) -> Result<(), CalcError> {
+    ctx.api
+        .swrite_i32(0, a + b)
+        .map_err(|_| CalcError::KelkError)
 }
 
-fn sub(_ctx: Context, a: i32, b: i32) -> Result<i32, CalcError> {
-    Ok(a - b)
+fn sub(ctx: Context, a: i32, b: i32) -> Result<(), CalcError> {
+    ctx.api
+        .swrite_i32(0, a - b)
+        .map_err(|_| CalcError::KelkError)
 }
 
-fn mul(_ctx: Context, a: i32, b: i32) -> Result<i32, CalcError> {
-    Ok(a * b)
+fn mul(ctx: Context, a: i32, b: i32) -> Result<(), CalcError> {
+    ctx.api
+        .swrite_i32(0, a * b)
+        .map_err(|_| CalcError::KelkError)
 }
 
-fn div(_ctx: Context, a: i32, b: i32) -> Result<i32, CalcError> {
+fn div(ctx: Context, a: i32, b: i32) -> Result<(), CalcError> {
     if b == 0 {
         return Err(CalcError::DivByZero);
     }
-    Ok(a / b)
+    ctx.api
+        .swrite_i32(0, a / b)
+        .map_err(|_| CalcError::KelkError)
 }
 
-// pub fn set_memory(ctx: Context, m: i32) -> Result<i32, CalcError> {
-//     let d = m.to_be_bytes();
-//     ctx.write_storage(0, d)?;
-//     Ok(m)
-// }
-
-// pub fn memory(ctx: Context) -> Result<i32, CalcError> {
-//     let d = ctx.read_storage(0, 4)?;
-//     let m = i32::from_be_bytes(d);
-//     Ok(m)
-// }
+fn query_result(ctx: Context) -> Result<i32, CalcError> {
+    ctx.api.sread_i32(0).map_err(|_| CalcError::KelkError)
+}
 
 /// The "instantiate" will be executed only once on instantiating the contract actor
 #[cfg(target_arch = "wasm32")]
@@ -50,22 +50,36 @@ mod __wasm_export_process_msg {
     }
 }
 
+#[cfg(target_arch = "wasm32")]
+mod __wasm_export_query {
+    #[no_mangle]
+    extern "C" fn query(msg_ptr: u64) -> u64 {
+        kelk_env::do_query(&super::query, msg_ptr)
+    }
+}
+
 // #[kelk_derive(instantiate)]
-pub fn instantiate(_ctx: Context) -> Result<Response, CalcError> {
-    Ok(Response { res: 0 })
+pub fn instantiate(_ctx: Context) -> Result<(), CalcError> {
+    Ok(())
 }
 
 /// _ctx: Context) The process_msg function is the main function of the *deployed* contract actor
 // #[kelk_derive(process_msg)]
-pub fn process_msg(ctx: Context, msg: CalcMsg) -> Result<Response, CalcError> {
-    let ans = match msg {
-        CalcMsg::Add { a, b } => add(ctx, a, b),
-        CalcMsg::Sub { a, b } => sub(ctx, a, b),
-        CalcMsg::Mul { a, b } => mul(ctx, a, b),
-        CalcMsg::Div { a, b } => div(ctx, a, b),
+pub fn process_msg(ctx: Context, msg: ProcMsg) -> Result<(), CalcError> {
+    match msg {
+        ProcMsg::Add { a, b } => add(ctx, a, b),
+        ProcMsg::Sub { a, b } => sub(ctx, a, b),
+        ProcMsg::Mul { a, b } => mul(ctx, a, b),
+        ProcMsg::Div { a, b } => div(ctx, a, b),
+    }
+}
+
+pub fn query(ctx: Context, msg: QueryMsg) -> Result<QueryRsp, CalcError> {
+    let res = match msg {
+        QueryMsg::LastResult => query_result(ctx),
     }?;
 
-    Ok(Response { res: ans })
+    Ok(QueryRsp::Result { res })
 }
 
 #[cfg(test)]