+++ examples/calculator/src/error.rs
@@ -3,5 +3,7 @@ use minicbor::Encode;
 #[derive(Clone, Debug, Encode)]
 pub enum CalcError {
     #[n(0)]
+    KelkError,
+    #[n(1)]
     DivByZero,
 }