+++ examples/calculator/src/message.rs
@@ -1,7 +1,7 @@
 use minicbor::{Decode, Encode};
 
 #[derive(Clone, Debug, Encode, Decode)]
-pub enum CalcMsg {
+pub enum ProcMsg {
     #[n(0)]
     Add {
         #[n(0)]
@@ -29,5 +29,20 @@ pub enum CalcMsg {
         a: i32,
         #[n(1)]
         b: i32,
+    },
+}
+
+#[derive(Clone, Debug, Encode, Decode)]
+pub enum QueryMsg {
+    #[n(0)]
+    LastResult,
+}
+
+#[derive(Clone, Debug, Encode, Decode)]
+pub enum QueryRsp {
+    #[n(0)]
+    Result {
+        #[n(0)]
+        res: i32,
     },
 }