+++ kelk/env/src/context.rs
@@ -2,17 +2,10 @@
 
 use crate::error::KelkError;
 use crate::params::*;
-use alloc::vec::Vec;
 
 /// `ContextAPI` provides the necessary APIs to interact with the Tanour.
 /// It can't be copied or cloned since it doesn't have Copy and Clone traits.
-pub trait ContextAPI {
-    /// Writes `data` into the storage file at the given offset
-    fn write_storage(&self, offset: u32, data: &[u8]) -> Result<(), KelkError>;
-
-    /// Writes `data` from the storage file at the given offset and length
-    fn read_storage(&self, offset: u32, length: u32) -> Result<Vec<u8>, KelkError>;
-
+pub trait ContextAPI: kelk_lib::storage::Storage {
     /// Gets parameters
     fn get_param(&self, param_id: i32) -> Result<ParamType, KelkError>;
 }