+++ kelk/env/src/error.rs
@@ -1,9 +1,6 @@
 //! Kelk errors.
 
+// TODO: delete it if not necessary !
+
 /// Kelk error types
-pub enum KelkError {
-    /// An issue occurred on writing into storage file
-    WriteStorageFailed,
-    /// if user try to access out of bound offset
-    StorageOutOfBound,
-}
+pub enum KelkError {}