+++ kelk/env/src/export.rs
@@ -25,18 +25,16 @@ extern "C" fn deallocate(ptr_u64: u64) {
     Pointer::from_u64(ptr_u64).deallocate();
 }
 
-/// TODO
+/// do_instantiate should be wrapped in an external "C" export,
+/// containing a contract-specific function as arg.
 pub fn do_instantiate<E: Encode>(instantiate_fn: &dyn Fn(Context) -> E) -> u32 {
     let ctx = make_context();
     instantiate_fn(ctx.as_ref());
     0
 }
 
-/// TODO: UPDATE MY COMMENT
-/// do_process_msg should be wrapped in an external "C" export, containing a contract-specific function as arg
-///
-/// - `M`: message type for request
-/// - `E`: error type for responses
+/// do_process_msg should be wrapped in an external "C" export,
+/// containing a contract-specific function as arg.
 pub fn do_process_msg<'a, D: Decode<'a>, E: Encode>(
     process_msg_fn: &dyn Fn(Context, D) -> E,
     msg_ptr: u64,
@@ -50,8 +48,23 @@ pub fn do_process_msg<'a, D: Decode<'a>, E: Encode>(
     result_to_ptr(res)
 }
 
+/// do_query should be wrapped in an external "C" export,
+/// containing a contract-specific function as arg.
+pub fn do_query<'a, D: Decode<'a>, E: Encode>(
+    query_fn: &dyn Fn(Context, D) -> E,
+    msg_ptr: u64,
+) -> u64 {
+    let ptr = Pointer::from_u64(msg_ptr);
+    let buf = unsafe { ptr.to_slice() };
+    let msg = minicbor::decode(buf).expect("Decoding failed");
+    let ctx = make_context();
+    let res = query_fn(ctx.as_ref(), msg);
+
+    result_to_ptr(res)
+}
+
 fn result_to_ptr<E: Encode>(res: E) -> u64 {
-    let mut vec = alloc::vec::Vec::new();
+    let mut vec = kelk_lib::alloc::vec::Vec::new();
     minicbor::encode(res, &mut vec).expect("Encoding failed");
 
     Pointer::release_buffer(vec).as_u64()