+++ kelk/env/src/memory.rs
@@ -1,4 +1,5 @@
-use alloc::vec::Vec;
+use kelk_lib::alloc;
+use kelk_lib::alloc::vec::Vec;
 
 /// defines a pointer to the allocated space in Wasm's linear memory.
 pub struct Pointer {