+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Defining Memory Pointer for managing memory allocation in WASM's linear memory.
 - Implementing `deallocate` API to free the allocated memory
+- Implementing Query API
+- Updating Storage API for primitive types
 
 ## Version 0.1.0